fix bug: SEQUOIADBMAINSTREAM-1441 在全组rebuild后，加入新的节点，新节点同步时发生core
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21475 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdMain.cpp
@@ -145,7 +145,7 @@ namespace engine
             {
                /// when rebuild, we can't move the dps to 0, because the new add
                /// node will sync from lsn 0
-               expectLSN.offset = ossAlign4( sizeof( dpsLogRecordHeader ) ) ;
+               expectLSN.offset = ossAlign4( (UINT32)sizeof( dpsLogRecordHeader ) ) ;
             }
             if ( DPS_INVALID_LSN_VERSION == expectLSN.version )
             {
